fix(ci): split the reminders out of MeetingStore, which the merge pushed over its limit
`Lint Check` failed on the merge commit with

    MeetingStore.swift:8:7: error: Type Body Length Violation:
    Class body should span 450 lines or less excluding comments and
    whitespace: currently spans 461 lines (type_body_length)

Neither branch broke it alone. This branch added the space and task
operations, #58 added the transcript and diarization ones, and merging
stacked both onto the same class body — eleven lines past the limit.

`MeetingStore+Reminders.swift` takes the action-item due dates and their
scheduled notifications, which is the store's only dependency on
`ReminderManager` and the one section that pairs a stored date with a
system notification. That is the split the file's own footer already
documents for eight other extensions, and it takes the body to 410.

Verified at the pinned tool versions CI uses rather than whatever is on
PATH: SwiftFormat 0.62.1 `--lint .` and SwiftLint 0.65.0
`lint --strict Logue/` both clean, both targets build, and the full
no-model suite passes — 1452 tests in 123 suites.
